+++ src/Utils.cpp
@@ -8,7 +8,7 @@ namespace sat::utils {
 // Random number generator
 // =============================================================================
 std::random_device RandomGen::rd;
-unsigned int RandomGen::initialSeed = RandomGen::rd();
+unsigned long RandomGen::initialSeed = RandomGen::rd();
 std::mt19937 RandomGen::randomGenerator(initialSeed);
 std::uniform_int_distribution<> RandomGen::randomBoolUD(0, 1);
 std::uniform_real_distribution<> RandomGen::randomReal01UD(0, 1);
